doc: Relocate Doxygen tools into doxygen_cxx subdirectory
We want the website's subdirectories to match the source tree's
subdirectories.

Also factor out a potential rule-of-three violation for build tag
named constants.

Also fix the Doxygen upstream website hyperlink to be correct.

Also fix assertion typo in doc/build.py.

Also remove a stray reference to Sphinx in the README.
