+++ doc/doxygen_cxx/Doxyfile_CXX.in
@@ -103,7 +103,7 @@ SHOW_USED_FILES        = YES
 SHOW_FILES             = YES
 SHOW_NAMESPACES        = YES
 FILE_VERSION_FILTER    =
-LAYOUT_FILE            = "@INPUT_ROOT@/drake/doc/DoxygenLayout.xml"
+LAYOUT_FILE            = "@INPUT_ROOT@/drake/doc/doxygen_cxx/DoxygenLayout.xml"
 CITE_BIB_FILES         =
 #---------------------------------------------------------------------------
 # Configuration options related to warning and progress messages
@@ -140,7 +140,7 @@ EXAMPLE_PATH           =
 EXAMPLE_PATTERNS       =
 EXAMPLE_RECURSIVE      = NO
 IMAGE_PATH             = "@INPUT_ROOT@/drake"
-INPUT_FILTER           = "@INPUT_ROOT@/drake/doc/system_doxygen.py"
+INPUT_FILTER           = "@INPUT_ROOT@/drake/doc/doxygen_cxx/system_doxygen.py"
 FILTER_PATTERNS        =
 FILTER_SOURCE_FILES    = NO
 FILTER_SOURCE_PATTERNS =
@@ -170,9 +170,9 @@ GENERATE_HTML          = YES
 HTML_OUTPUT            = html
 HTML_FILE_EXTENSION    = .html
 HTML_HEADER            =
-HTML_FOOTER            = "@INPUT_ROOT@/drake/doc/footer.html"
+HTML_FOOTER            = "@INPUT_ROOT@/drake/doc/doxygen_cxx/footer.html"
 HTML_STYLESHEET        =
-HTML_EXTRA_STYLESHEET  = "@INPUT_ROOT@/drake/doc/doxygen_extra.css"
+HTML_EXTRA_STYLESHEET  = "@INPUT_ROOT@/drake/doc/doxygen_cxx/doxygen_extra.css"
 HTML_EXTRA_FILES       =
 HTML_COLORSTYLE_HUE    = 220
 HTML_COLORSTYLE_SAT    = 100
