+++ doc/doxygen_cxx/footer.html
