+++ doc/pydrake/BUILD.bazel
@@ -7,33 +7,17 @@ load(
     "drake_py_binary",
     "drake_py_library",
 )
+load(
+    "//doc:defs.bzl",
+    "DEFAULT_BINARY_TAGS",
+    "DEFAULT_TEST_TAGS",
+)
 load("//bindings/pydrake:pydrake.bzl", "add_lint_tests_pydrake")
-
-# Unless `setup/ubuntu/install_prereqs.sh --with-doc-only` has been run, most
-# targets in this package will fail to build, so by default we'll disable them.
-#
-# A developer will have to explicitly opt-in in order to build these.
-_DEFAULT_BINARY_TAGS = [
-    "manual",
-]
-
-# Unless `setup/ubuntu/install_prereqs.sh --with-doc-only` has been run, most
-# tests in this package will fail to pass, so by default we'll disable them.
-#
-# A developer will have to explicitly opt-in in order to test these.
-_DEFAULT_TEST_TAGS = [
-    "manual",
-    # Some (currently disabled) Sphinx extensions try to reach out to the
-    # network; we should fail-fast if someone tries to turn them on.  In
-    # general, none our documentation tools should hit the internet, but
-    # their ecosystems might be doing so without us being aware.
-    "block-network",
-]
 
 drake_py_library(
     name = "sphinx_base",
     srcs = ["sphinx_base.py"],
-    tags = _DEFAULT_BINARY_TAGS,
+    tags = DEFAULT_BINARY_TAGS,
     deps = ["@rules_python//python/runfiles"],
 )
 
@@ -45,7 +29,7 @@ drake_py_library(
         ":sphinx_base",
         "//bindings/pydrake/common:cpp_template_py",
         "//bindings/pydrake/common:deprecation_py",
-        "//doc:system_doxygen",
+        "//doc/doxygen_cxx:system_doxygen",
     ],
 )
 
@@ -61,10 +45,10 @@ drake_py_binary(
     ],
     imports = ["."],
     main = "gen_sphinx.py",
-    tags = _DEFAULT_BINARY_TAGS,
+    tags = DEFAULT_BINARY_TAGS,
     test_rule_args = ["--out_dir=<test>"],
     test_rule_size = "medium",
-    test_rule_tags = _DEFAULT_TEST_TAGS,
+    test_rule_tags = DEFAULT_TEST_TAGS,
     visibility = ["//doc:__pkg__"],
     deps = [
         ":pydrake_sphinx_extension_py",
@@ -77,7 +61,7 @@ drake_py_binary(
     name = "serve_sphinx",
     srcs = ["serve_sphinx.py"],
     data = [":gen_sphinx"],
-    tags = _DEFAULT_BINARY_TAGS,
+    tags = DEFAULT_BINARY_TAGS,
     visibility = ["//doc:__pkg__"],
     deps = [":sphinx_base"],
 )
