+++ tools/workspace/pybind11/BUILD.bazel
@@ -50,7 +50,7 @@ py_library(
     name = "mkdoc_comment_py",
     srcs = ["mkdoc_comment.py"],
     visibility = ["//visibility:public"],
-    deps = ["//doc:system_doxygen"],
+    deps = ["//doc/doxygen_cxx:system_doxygen"],
 )
 
 drake_py_unittest(
